pyproject.toml
[project]
name = "dial"
version = "0.1.0"
description = "A Python library for procedural and declarative generation of analog clock faces"
readme = "README.md"
authors = [
    { name = "Vyoman Jain", email = "75789232+VyoJ@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "Pillow",
    "numpy",
    "typer",
]

[project.scripts]
dial = "dial.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "mkdocs>=1.4.0",
    "mkdocs-material>=8.0.0",
    "mkdocstrings[python]>=0.20.0"
]

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.21",
    "mkdocstrings[python]>=0.30.1",
    "pytest>=8.4.2",
]
